apps/backend/requirements.txt
# Core Django
Django==5.2.10
asgiref==3.11.0
sqlparse==0.5.5
typing_extensions==4.15.0

# Django REST
djangorestframework==3.16.1
djangorestframework-simplejwt==5.4.0
django-filter==24.2
django-cors-headers==4.9.0

# Database
psycopg==3.3.2

# Environment
python-dotenv==1.2.1

# Internal shared packages (monorepo)
-e packages/snowflake-id-lib/
